Implement feature-based restructure with barrel exports and search feature
