+++ src/components/common/index.js
@@ -1,0 +1,6 @@
+// Common components barrel exports
+export { default as Layout } from './Layout.jsx';
+export { default as LoadingSpinner } from './LoadingSpinner.jsx';
+export { default as Section } from './Section.jsx';
+export { default as StateRenderer } from './StateRenderer.jsx';
+export { default as withState } from './withState.jsx';
