+++ src/features/items/index.js
@@ -1,0 +1,14 @@
+// Items feature barrel exports
+
+// Components
+export { default as ItemCard } from './components/ItemCard.jsx';
+export { default as ItemForm } from './components/ItemForm.jsx';
+
+// Hooks
+export { default as useItems } from './hooks/useItems.js';
+
+// Services
+export * from './services/itemApi.js';
+
+// Types
+export * from './types/index.js';
