+++ src/features/search/index.js
@@ -1,0 +1,10 @@
+// Search feature barrel exports
+
+// Components
+export { default as SearchBar } from './components/SearchBar.jsx';
+export { default as SearchHero } from './components/SearchHero.jsx';
+export { default as SearchResults } from './components/SearchResults.jsx';
+export { default as SearchSuggestions } from './components/SearchSuggestions.jsx';
+
+// Hooks
+export { default as useSearch } from './hooks/useSearch.js';
