+++ src/hooks/index.js
@@ -1,0 +1,3 @@
+// Global hooks barrel exports
+export { default as useHealthCheck } from './useHealthCheck.js';
+export { default as useProfile } from './useProfile.js';
