+++ src/services/index.js
@@ -1,0 +1,2 @@
+// Services barrel exports
+export { default, apiService } from './api.js';
