+++ src/utils/index.js
@@ -1,0 +1,3 @@
+// Utils barrel exports
+export * from './constants.js';
+export * from './textUtils.js';
